清理冗余代码和过时模块，重构延迟探测与状态同步

+++ app/src/main/java/com/kunk/singbox/utils/parser/NodeLinkParserTypes.kt
@@ -1,11 +1,8 @@
 ﻿package com.kunk.singbox.utils.parser
 
-/**
- */
-
 data class NodeLinkParserWebSocketPathConfig(
     val path: String,
-    val maxEarlyData: Int?,
+    val maxEarlyData: Long?,
     val earlyDataHeaderName: String?
 )
